Bloom suites before debriefing
Fixes an issue where opening a suit mid-run would cause the tests that had already run to be bloomed as queued (regardless of existing results) and then removed when the suite finished running. Not only blooming then pre-debrief is the easiest way to ensure we have the proper models to work with, but it also fixes the issue of running suites not being marked as such.

Also changed the parsing of statuses to mark parents with skipped tests as "incomplete" instead of "partial".