Exclude hidden part of a suite's path from keyword filtering
Parsing of keyword for filtering purposes will now remove the local or remote repository path from the string, and matches are made by comparing only the visible/relative path of a suite. Hopeffuly that means matches should be a bit more logical and we can pave the way for better match feedback.