Improvements in keyword filtering
On the last update we started processing the keyword to remove the repositories paths from it, then comparing to the existing relative suites to find matches. This approach was limited to either searching for the full path or searching for the relative portions. Unfortunately, it did not work at all when the keyword contained a fraction of the full/hidden repository path, which seemed a bit counter-intuitive, especially if tests path were configured (i.e. Sublime can let you copy a file's path, but it will be relative to the repositorie's root folder, not the tests folder).

A new approach means we're searching for each directory in the repository's path (including configured tests path) and, if present, we'll remove that from the keyword before attempting to find matches. This is much more lenient and feels sufficiently intuitive.