Fix filename truncation with relative paths
Since truncated filename's are RTL text, when they were preceded by "/" or ".." or any combination of those, that portion of the path would be placed at the end of the string, mangling its meaning. This is because of how the browser handles special characters in RTL, but can be fixed by preceding the string with a Unicode Left-to-Right mark, which makes the browser understand that the whole sequence should be treated as LTR text, rendered in RTL for overflowing purposes.