Added migration logic
We can now do migrations on the app's state, which are run before the window is created, if necessary (runs upon instantiating state in the main process). It aims to make migration as painless as possible, though it could still use some polish.

We decided, though, not to use it after all. The first/last test properties are fairly harmless and should be computed automatically when browsing through the tests, so having a migration just for those felt unnecessarily risky. It would not even be necessary save for the blooming/withering technique, which only creates tests when needed. Alas, the worst-case scenario for lacking a migration seems to be slightly skewed "first seen" dates, which I believe is compromise worth not risking compromising user data.