Added errorQueued method for nuggets
Fixes an issue (with potentially all frameworks, but which only affected PHPUnit) in which suites would get stuck in "running" status if running a subset of tests in which one or more had been removed from the framework. Previously we would only act upon the unexpected amount of tests run if all within subset had been removed and lacked a way to granularly mark tests within the modified suite as having errored out.