Fixed lingering "queued" status in test results
When interrupting a test run, remaining queued tests are made idle again. But we neglected to update their result objects, which could make them appear as "queued" again if the status needed to be parsed again while they were withered (e.g. when reloading the framework without running), in turn making their parent adopt a "running" status even though no run was in progress.

Also includes better logging of the project to the console from the app menu (when in development).