Add support for PHPUnit 9+
In PHPUnit 9+, the ResultPrinter class became an interface, so we need to override Lode's default Printer class in PHPUnit 7 & 8, which will reference the previous default printer class to prevent it from throwing an error. The new version of the PHPUnit reporter going forward will extend a new, local "default printer", which we can override at will going forward.

Fixes #48.